Fix scalar/wedge consistency with error-weighted binning
