+++ src/drtsans/tof/eqsans/correction_api.py
@@ -5,10 +5,8 @@ from collections import namedtuple
 from typing import List, Union
 
 from drtsans.iq import bin_all  # noqa E402
-from drtsans.tof.eqsans.incoherence_correction_1d import (
+from drtsans.tof.eqsans.incoherence_correction import (
     CorrectedI1D,
-)
-from drtsans.tof.eqsans.incoherence_correction_2d import (
     CorrectedIQ2D,
 )
